requirements.txt
altgraph==0.16.1
astroid==2.2.5
beautifulsoup4==4.8.0
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
et-xmlfile==1.0.1
Flask==1.1.1
future==0.17.1
gunicorn==19.9.0
idna==2.8
inflect==2.1.0
isort==4.3.21
itsdangerous==1.1.0
jaraco.itertools==4.4.2
jdcal==1.4.1
Jinja2==2.10.1
lazy-object-proxy==1.4.2
line-bot-sdk==1.14.0
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==7.2.0
MouseInfo==0.0.4
numpy==1.17.2
opencv-python==4.1.1.26
openpyxl==2.6.3
pefile==2019.4.18
Pillow==6.1.0
PyAutoGUI==0.9.47
PyGetWindow==0.0.7
PyInstaller==3.5
pylint==2.3.1
PyMsgBox==1.0.7
pyperclip==1.7.0
PyRect==0.1.4
PyScreeze==0.1.22
pytesseract==0.3.0
PyTweening==1.0.3
pywin32-ctypes==0.2.0
requests==2.22.0
selenium==3.141.0
six==1.12.0
songline==0.0.1
soupsieve==1.9.3
tesseract==0.1.3
typed-ast==1.4.0
uncleengineer==0.2
urllib3==1.25.3
virtualenv==16.7.5
Werkzeug==0.15.6
wikipedia==1.4.0
wolframalpha==3.0.1
wrapt==1.11.2
xmltodict==0.12.0
